kubernetes/requirements.txt
Flask==3.0.0
PyYAML==6.0.1
kubernetes==18.20.0